server/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
esprima==4.0.1
fastparquet==2024.5.0
fasttext-wheel==0.9.2
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
ml-dtypes==0.4.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
soupsieve==2.5
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.14.1
